requirements.txt
celery==5.3.4
django-cors-headers==4.3.1
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
pymongo==4.6.0
python-dotenv==1.0.0
redis==5.0.1
setuptools==69.0.0
textblob==0.17.1